Review requested changes
