+++ ui/pages/send/gas-display/gas-display.js
@@ -1,5 +1,5 @@
 import React, { useContext, useState } from 'react';
-import { useSelector } from 'react-redux';
+import { useSelector, useDispatch } from 'react-redux';
 import PropTypes from 'prop-types';
 import { I18nContext } from '../../../contexts/i18n';
 import { useGasFeeContext } from '../../../contexts/gasFee';
@@ -7,9 +7,8 @@ import { PRIMARY, SECONDARY } from '../../../helpers/constants/common';
 import { hexWEIToDecGWEI } from '../../../../shared/lib/transactions-controller-utils';
 import UserPreferencedCurrencyDisplay from '../../../components/app/user-preferenced-currency-display';
 import GasTiming from '../../../components/app/gas-timing';
-import InfoTooltip from '../../../components/ui/info-tooltip/info-tooltip';
-import LedgerInstructionField from '../../../components/app/ledger-instruction-field';
-import Typography from '../../../components/ui/typography/typography';
+import InfoTooltip from '../../../components/ui/info-tooltip';
+import Typography from '../../../components/ui/typography';
 import Button from '../../../components/ui/button';
 import Box from '../../../components/ui/box';
 import {
@@ -18,144 +17,125 @@ import {
   FLEX_DIRECTION,
   BLOCK_SIZES,
 } from '../../../helpers/constants/design-system';
+import {
+  ERC1155,
+  ERC20,
+  ERC721,
+} from '../../../../shared/constants/transaction';
 import LoadingHeartBeat from '../../../components/ui/loading-heartbeat';
 import TransactionDetailItem from '../../../components/app/transaction-detail-item';
 import { NETWORK_TO_NAME_MAP } from '../../../../shared/constants/network';
 import TransactionDetail from '../../../components/app/transaction-detail';
 import ActionableMessage from '../../../components/ui/actionable-message';
 import DepositPopover from '../../../components/app/deposit-popover';
-import { getProvider } from '../../../selectors';
-import { INSUFFICIENT_TOKENS_ERROR } from '../send.constants';
+import {
+  getProvider,
+  getPreferences,
+  getIsBuyableChain,
+  transactionFeeSelector,
+  getIsMainnet,
+} from '../../../selectors';
 
-const renderHeartBeatIfNotInTest = () =>
-  process.env.IN_TEST ? null : <LoadingHeartBeat />;
-export default function GasDisplay({
-  draftTransaction,
-  detailText,
-  detailTotal,
-  maxAmount,
-  useNativeCurrencyAsPrimaryCurrency,
-  maxFeePerGas,
-  maxPriorityFeePerGas,
-  isMainnet,
-  showLedgerSteps,
-  isBuyableChain,
-  hexMinimumTransactionFee,
-  hexMaximumTransactionFee,
-  nativeCurrency,
-  chainId,
-  showAccountDetails,
-  gasError,
-}) {
+import {
+  hexWEIToDecETH,
+  addHexes,
+} from '../../../helpers/utils/conversions.util';
+import { INSUFFICIENT_TOKENS_ERROR } from '../send.constants';
+import { getCurrentDraftTransaction } from '../../../ducks/send';
+import { showModal } from '../../../store/actions';
+
+export default function GasDisplay({ gasError }) {
   const t = useContext(I18nContext);
+  const dispatch = useDispatch();
   const { estimateUsed, supportsEIP1559V2 } = useGasFeeContext();
   const [showDepositPopover, setShowDepositPopover] = useState(false);
   const currentProvider = useSelector(getProvider);
+  const isMainnet = useSelector(getIsMainnet);
+  const isBuyableChain = useSelector(getIsBuyableChain);
+  const draftTransaction = useSelector(getCurrentDraftTransaction);
+  const { useNativeCurrencyAsPrimaryCurrency } = useSelector(getPreferences);
+  const { nativeCurrency, provider, unapprovedTxs } = useSelector(
+    (state) => state.metamask,
+  );
+  const { chainId } = provider;
   const networkName = NETWORK_TO_NAME_MAP[chainId];
   const isInsufficientTokenError =
     draftTransaction?.amount.error === INSUFFICIENT_TOKENS_ERROR;
+  const editingTransaction = unapprovedTxs[draftTransaction.id];
 
-  const renderGasDetailsItem = () => {
-    return (
-      <TransactionDetailItem
-        key="gas-item"
-        detailTitle={
-          <>
-            {t('transactionDetailGasHeading')}
-            <InfoTooltip
-              contentText={
-                <>
-                  <p>
-                    {t('transactionDetailGasTooltipIntro', [
-                      isMainnet ? t('networkNameEthereum') : '',
-                    ])}
-                  </p>
-                  <p>{t('transactionDetailGasTooltipExplanation')}</p>
-                  <p>
-                    <a
-                      href="https://community.metamask.io/t/what-is-gas-why-do-transactions-take-so-long/3172"
-                      target="_blank"
-                      rel="noopener noreferrer"
-                    >
-                      {t('transactionDetailGasTooltipConversion')}
-                    </a>
-                  </p>
-                </>
-              }
-              position="top"
-            >
-              <i className="fa fa-info-circle" />
-            </InfoTooltip>
-          </>
-        }
-        detailText={
-          <Box
-            height={BLOCK_SIZES.MAX}
-            display={DISPLAY.FLEX}
-            flexDirection={FLEX_DIRECTION.COLUMN}
-            className="gas-display__currency-container"
-          >
-            {renderHeartBeatIfNotInTest()}
-            <UserPreferencedCurrencyDisplay
-              type={SECONDARY}
-              value={hexMinimumTransactionFee}
-              hideLabel={Boolean(useNativeCurrencyAsPrimaryCurrency)}
-            />
-          </Box>
-        }
-        detailTotal={
-          <Box
-            height={BLOCK_SIZES.MAX}
-            display={DISPLAY.FLEX}
-            flexDirection={FLEX_DIRECTION.COLUMN}
-            className="gas-display__currency-container"
-          >
-            {renderHeartBeatIfNotInTest()}
-            <UserPreferencedCurrencyDisplay
-              type={PRIMARY}
-              value={hexMinimumTransactionFee}
-              hideLabel={!useNativeCurrencyAsPrimaryCurrency}
-              numberOfDecimals={6}
-            />
-          </Box>
-        }
-        subText={
-          <>
-            <strong key="editGasSubTextFeeLabel">
-              {t('editGasSubTextFeeLabel')}
-            </strong>
-            <Box
-              height={BLOCK_SIZES.MAX}
-              display={DISPLAY.FLEX}
-              flexDirection={FLEX_DIRECTION.COLUMN}
-              className="gas-display__currency-container"
-            >
-              {renderHeartBeatIfNotInTest()}
-              <UserPreferencedCurrencyDisplay
-                key="editGasSubTextFeeAmount"
-                type={PRIMARY}
-                value={hexMaximumTransactionFee}
-                hideLabel={!useNativeCurrencyAsPrimaryCurrency}
-              />
-            </Box>
-          </>
-        }
-        subTitle={
-          <>
-            <GasTiming
-              maxPriorityFeePerGas={hexWEIToDecGWEI(
-                maxPriorityFeePerGas ||
-                  draftTransaction.gas.maxPriorityFeePerGas,
-              )}
-              maxFeePerGas={hexWEIToDecGWEI(
-                maxFeePerGas || draftTransaction.gas.maxFeePerGas,
-              )}
-            />
-          </>
-        }
+  const txData = {
+    txParams: {
+      gasPrice: draftTransaction.gas?.gasPrice,
+      gas: editingTransaction?.userEditedGasLimit
+        ? editingTransaction?.txParams?.gas
+        : draftTransaction.gas?.gasLimit,
+      maxFeePerGas: draftTransaction.gas?.maxFeePerGas,
+      maxPriorityFeePerGas: draftTransaction.gas?.maxPriorityFeePerGas,
+      value: draftTransaction.amount?.value,
+      type: draftTransaction.transactionType,
+    },
+    userFeeLevel: editingTransaction?.userFeeLevel,
+  };
+
+  const {
+    hexMinimumTransactionFee,
+    hexMaximumTransactionFee,
+    hexTransactionTotal,
+  } = useSelector((state) => transactionFeeSelector(state, txData));
+
+  let title;
+  if (
+    draftTransaction?.asset.details?.standard === ERC721 ||
+    draftTransaction?.asset.details?.standard === ERC1155
+  ) {
+    title = draftTransaction?.asset.details?.name;
+  } else if (draftTransaction?.asset.details?.standard === ERC20) {
+    title = `${hexWEIToDecETH(draftTransaction.amount.value)} ${
+      draftTransaction?.asset.details?.symbol
+    }`;
+  }
+
+  const ethTransactionTotalMaxAmount = Number(
+    hexWEIToDecETH(hexMaximumTransactionFee),
+  );
+
+  const primaryTotalTextOverrideMaxAmount = `${title} + ${ethTransactionTotalMaxAmount} ${nativeCurrency}`;
+
+  let detailTotal, maxAmount;
+
+  if (draftTransaction?.asset.type === 'NATIVE') {
+    detailTotal = (
+      <Box
+        height={BLOCK_SIZES.MAX}
+        display={DISPLAY.FLEX}
+        flexDirection={FLEX_DIRECTION.COLUMN}
+        className="gas-display__total-value"
+      >
+        <LoadingHeartBeat estimateUsed={txData?.userFeeLevel} />
+        <UserPreferencedCurrencyDisplay
+          type={PRIMARY}
+          key="total-detail-value"
+          value={hexTransactionTotal}
+          hideLabel={!useNativeCurrencyAsPrimaryCurrency}
+        />
+      </Box>
+    );
+    maxAmount = (
+      <UserPreferencedCurrencyDisplay
+        type={PRIMARY}
+        key="total-max-amount"
+        value={addHexes(
+          draftTransaction.amount.value,
+          hexMaximumTransactionFee,
+        )}
+        hideLabel={!useNativeCurrencyAsPrimaryCurrency}
       />
     );
-  };
+  } else if (useNativeCurrencyAsPrimaryCurrency) {
+    detailTotal = primaryTotalTextOverrideMaxAmount;
+    maxAmount = primaryTotalTextOverrideMaxAmount;
+  }
+
   return (
     <>
       {showDepositPopover && (
@@ -165,12 +145,124 @@ export default function GasDisplay({
         <TransactionDetail
           userAcknowledgedGasMissing={false}
           rows={[
-            renderGasDetailsItem(),
+            <TransactionDetailItem
+              key="gas-item"
+              detailTitle={
+                <>
+                  {t('transactionDetailGasHeading')}
+                  <InfoTooltip
+                    contentText={
+                      <>
+                        <p>
+                          {t('transactionDetailGasTooltipIntro', [
+                            isMainnet ? t('networkNameEthereum') : '',
+                          ])}
+                        </p>
+                        <p>{t('transactionDetailGasTooltipExplanation')}</p>
+                        <p>
+                          <a
+                            href="https://community.metamask.io/t/what-is-gas-why-do-transactions-take-so-long/3172"
+                            target="_blank"
+                            rel="noopener noreferrer"
+                          >
+                            {t('transactionDetailGasTooltipConversion')}
+                          </a>
+                        </p>
+                      </>
+                    }
+                    position="right"
+                  >
+                    <i className="fa fa-info-circle" />
+                  </InfoTooltip>
+                </>
+              }
+              detailText={
+                <Box
+                  height={BLOCK_SIZES.MAX}
+                  display={DISPLAY.FLEX}
+                  flexDirection={FLEX_DIRECTION.COLUMN}
+                  className="gas-display__currency-container"
+                >
+                  <LoadingHeartBeat estimateUsed={txData?.userFeeLevel} />
+                  <UserPreferencedCurrencyDisplay
+                    type={SECONDARY}
+                    value={hexMinimumTransactionFee}
+                    hideLabel={Boolean(useNativeCurrencyAsPrimaryCurrency)}
+                  />
+                </Box>
+              }
+              detailTotal={
+                <Box
+                  height={BLOCK_SIZES.MAX}
+                  display={DISPLAY.FLEX}
+                  flexDirection={FLEX_DIRECTION.COLUMN}
+                  className="gas-display__currency-container"
+                >
+                  <LoadingHeartBeat estimateUsed={txData?.userFeeLevel} />
+                  <UserPreferencedCurrencyDisplay
+                    type={PRIMARY}
+                    value={hexMinimumTransactionFee}
+                    hideLabel={!useNativeCurrencyAsPrimaryCurrency}
+                    numberOfDecimals={6}
+                  />
+                </Box>
+              }
+              subText={
+                <>
+                  <strong key="editGasSubTextFeeLabel">
+                    {t('editGasSubTextFeeLabel')}
+                  </strong>
+                  <Box
+                    height={BLOCK_SIZES.MAX}
+                    display={DISPLAY.FLEX}
+                    flexDirection={FLEX_DIRECTION.COLUMN}
+                    className="gas-display__currency-container"
+                  >
+                    <LoadingHeartBeat
+                      estimateUsed={txData?.userFeeLevel ?? estimateUsed}
+                    />
+                    <UserPreferencedCurrencyDisplay
+                      key="editGasSubTextFeeAmount"
+                      type={PRIMARY}
+                      value={hexMaximumTransactionFee}
+                      hideLabel={!useNativeCurrencyAsPrimaryCurrency}
+                    />
+                  </Box>
+                </>
+              }
+              subTitle={
+                <>
+                  <GasTiming
+                    maxPriorityFeePerGas={hexWEIToDecGWEI(
+                      draftTransaction.gas.maxPriorityFeePerGas,
+                    )}
+                    maxFeePerGas={hexWEIToDecGWEI(
+                      draftTransaction.gas.maxFeePerGas,
+                    )}
+                  />
+                </>
+              }
+            />,
             !supportsEIP1559V2 && (gasError || isInsufficientTokenError) && (
               <TransactionDetailItem
                 key="total-item"
                 detailTitle={t('total')}
-                detailText={detailText}
+                detailText={
+                  <Box
+                    height={BLOCK_SIZES.MAX}
+                    display={DISPLAY.FLEX}
+                    flexDirection={FLEX_DIRECTION.COLUMN}
+                    className="gas-display__total-value"
+                  >
+                    <LoadingHeartBeat estimateUsed={txData?.userFeeLevel} />
+                    <UserPreferencedCurrencyDisplay
+                      type={SECONDARY}
+                      key="total-detail-text"
+                      value={hexTransactionTotal}
+                      hideLabel={Boolean(useNativeCurrencyAsPrimaryCurrency)}
+                    />
+                  </Box>
+                }
                 detailTotal={detailTotal}
                 subTitle={t('transactionDetailGasTotalSubtitle')}
                 subText={
@@ -181,9 +273,7 @@ export default function GasDisplay({
                     className="gas-display__total-amount"
                   >
                     <LoadingHeartBeat
-                      estimateUsed={
-                        draftTransaction?.userFeeLevel ?? estimateUsed
-                      }
+                      estimateUsed={txData?.userFeeLevel ?? estimateUsed}
                     />
                     <strong key="editGasSubTextAmountLabel">
                       {t('editGasSubTextAmountLabel')}
@@ -195,13 +285,6 @@ export default function GasDisplay({
             ),
           ]}
         />
-        {showLedgerSteps ? (
-          <LedgerInstructionField
-            showDataInstruction={Boolean(
-              draftTransaction.userInputHexData ?? undefined,
-            )}
-          />
-        ) : null}
       </Box>
       {!supportsEIP1559V2 && (gasError || isInsufficientTokenError) && (
         <Box className="gas-display__warning-message">
@@ -232,7 +315,9 @@ export default function GasDisplay({
                       <Button
                         type="inline"
                         className="gas-display__link"
-                        onClick={showAccountDetails}
+                        onClick={() =>
+                          dispatch(showModal({ name: 'ACCOUNT_DETAILS' }))
+                        }
                         key="receive-button"
                       >
                         {t('deposit')}
@@ -251,7 +336,9 @@ export default function GasDisplay({
                       <Button
                         type="inline"
                         className="gas-display__link"
-                        onClick={showAccountDetails}
+                        onClick={() =>
+                          dispatch(showModal({ name: 'ACCOUNT_DETAILS' }))
+                        }
                         key="receive-button"
                       >
                         {t('deposit')}
@@ -271,20 +358,5 @@ export default function GasDisplay({
   );
 }
 GasDisplay.propTypes = {
-  draftTransaction: PropTypes.object,
-  detailText: PropTypes.oneOfType([PropTypes.string, PropTypes.node]),
-  detailTotal: PropTypes.oneOfType([PropTypes.string, PropTypes.node]),
-  maxAmount: PropTypes.oneOfType([PropTypes.string, PropTypes.node]),
-  useNativeCurrencyAsPrimaryCurrency: PropTypes.bool,
-  maxFeePerGas: PropTypes.string,
-  maxPriorityFeePerGas: PropTypes.string,
-  isMainnet: PropTypes.bool,
-  showLedgerSteps: PropTypes.bool,
-  isBuyableChain: PropTypes.bool,
-  hexMinimumTransactionFee: PropTypes.string,
-  hexMaximumTransactionFee: PropTypes.string,
-  nativeCurrency: PropTypes.string,
-  chainId: PropTypes.string,
-  showAccountDetails: PropTypes.func,
   gasError: PropTypes.string,
 };
