+++ ui/pages/send/send-content/send-content.container.js
@@ -5,10 +5,6 @@ import {
   getIsEthGasPriceFetched,
   getNoGasPriceFetched,
   checkNetworkOrAccountNotSupports1559,
-  getPreferences,
-  getIsBuyableChain,
-  transactionFeeSelector,
-  getUseNonceField,
 } from '../../../selectors';
 import {
   getIsBalanceInsufficient,
@@ -18,11 +14,7 @@ import {
   getRecipient,
   acknowledgeRecipientWarning,
   getRecipientWarningAcknowledgement,
-  getCurrentDraftTransaction,
 } from '../../../ducks/send';
-
-import { showModal } from '../../../store/actions';
-
 import SendContent from './send-content.component';
 
 function mapStateToProps(state) {
@@ -31,42 +23,6 @@ function mapStateToProps(state) {
   const recipient = getRecipient(state);
   const recipientWarningAcknowledged =
     getRecipientWarningAcknowledgement(state);
-  const isBuyableChain = getIsBuyableChain(state);
-  const draftTransaction = getCurrentDraftTransaction(state);
-
-  const { currentCurrency, nativeCurrency, provider, unapprovedTxs } =
-    state.metamask;
-  const { chainId } = provider;
-  const editingTransaction = unapprovedTxs[draftTransaction.id];
-  const txData = {
-    txParams: {
-      gasPrice:
-        editingTransaction?.txParams?.gasPrice ??
-        draftTransaction.gas?.gasPrice,
-      gas: editingTransaction?.userEditedGasLimit
-        ? editingTransaction?.txParams?.gas
-        : draftTransaction.gas?.gasLimit,
-      maxFeePerGas:
-        editingTransaction?.txParams?.maxFeePerGas ??
-        draftTransaction.gas?.maxFeePerGas,
-      maxPriorityFeePerGas:
-        editingTransaction?.txParams?.maxPriorityFeePerGas ??
-        draftTransaction.gas?.maxPriorityFeePerGas,
-      value: draftTransaction.amount?.value,
-      type:
-        editingTransaction?.txParams?.type ?? draftTransaction.transactionType,
-    },
-    userFeeLevel: editingTransaction?.userFeeLevel,
-  };
-
-  const {
-    hexTransactionAmount,
-    hexMinimumTransactionFee,
-    hexMaximumTransactionFee,
-    hexTransactionTotal,
-  } = transactionFeeSelector(state, txData);
-
-  const { useNativeCurrencyAsPrimaryCurrency } = getPreferences(state);
 
   return {
     isOwnedAccount: Boolean(
@@ -83,27 +39,15 @@ function mapStateToProps(state) {
     getIsBalanceInsufficient: getIsBalanceInsufficient(state),
     asset: getSendAsset(state),
     assetError: getAssetError(state),
-    useNonceField: getUseNonceField(state),
-    draftTransaction,
-    hexMaximumTransactionFee,
-    hexMinimumTransactionFee,
-    hexTransactionTotal,
-    hexTransactionAmount,
-    currentCurrency,
-    nativeCurrency,
-    useNativeCurrencyAsPrimaryCurrency,
-    isBuyableChain,
-    chainId,
     recipient,
     recipientWarningAcknowledged,
   };
 }
 
-const mapDispatchToProps = (dispatch) => {
+function mapDispatchToProps(dispatch) {
   return {
-    showAccountDetails: () => dispatch(showModal({ name: 'ACCOUNT_DETAILS' })),
     acknowledgeRecipientWarning: () => dispatch(acknowledgeRecipientWarning()),
   };
-};
+}
 
 export default connect(mapStateToProps, mapDispatchToProps)(SendContent);
